Switch to dynamic space-time diagram
This is for future simulation purposes.

I will note that I prefer a static SVG for these non-dynamic cases as
the document, on loading, will jump around as the components are loaded.
This is probably a common issue with web components so there should be a
more generic fix than simply switching to static.