Tweak mobile view
When viewing on mobile, the page was cutting off the left of the text,
making it hard to read the start of a sentence! We've hopefully fixed
that but changing to a grid view.

Also, we shrink the size of the text when the screen is too narrow as I
think the limit "math[display='block']" element is causing the width to
overflow, resulting in scrolling being needed to view the rest of a
sentence.