Set hearbeat time to period
Technically, it is possible that a message isn't sent exactly every tau
seconds even if we set the heartbeat time to the maximum. In the case
where an animation frame is too long, we might be a few milliseconds
later than tau seconds.

However, for our simulation purposes, a few milliseconds doesn't really
matter; the main point is highlight the periodic sending of messages to
neighbours.